+++ lib/rules/no-top-level-variables.ts
@@ -1,9 +1,20 @@
 // SPDX-License-Identifier: ISC
 
 import type {Rule} from 'eslint';
-import type {Expression, CallExpression, VariableDeclarator} from 'estree';
+import type {
+  Declaration,
+  Expression,
+  CallExpression,
+  VariableDeclaration,
+  VariableDeclarator
+} from 'estree';
 
 import {isTopLevel} from '../helpers';
+
+type Options = {
+  readonly constAllowed: ReadonlyArray<string>;
+  readonly kind: ReadonlyArray<string>;
+};
 
 const violationMessage = 'Variables at the top level are not allowed';
 
@@ -16,6 +27,59 @@ const kindValues = ['const', 'let', 'var'];
 
 function isRequireCall(node: CallExpression): boolean {
   return node.callee.type === 'Identifier' && node.callee.name === 'require';
+}
+
+function checker(
+  context: Rule.RuleContext,
+  options: Options,
+  node: VariableDeclaration
+) {
+  if (!Array.from(options.kind).includes(node.kind)) {
+    return;
+  }
+
+  let allowedDeclaration: (declaration: VariableDeclarator) => boolean;
+  if (node.kind === 'const') {
+    const isArrowFunctionAllowed = options.constAllowed.includes(
+      'ArrowFunctionExpression'
+    );
+    const isLiteralAllowed = options.constAllowed.includes('Literal');
+    const isMemberExpressionAllowed =
+      options.constAllowed.includes('MemberExpression');
+
+    allowedDeclaration = (declaration) => {
+      // type-coverage:ignore-next-line
+      switch ((declaration.init as Expression).type) {
+        case 'ArrowFunctionExpression':
+          return isArrowFunctionAllowed;
+        case 'CallExpression': {
+          // type-coverage:ignore-next-line
+          return isRequireCall(declaration.init as CallExpression);
+        }
+        case 'Identifier':
+          return true;
+        case 'Literal':
+          return isLiteralAllowed;
+        case 'MemberExpression':
+          return isMemberExpressionAllowed;
+        default:
+          return false;
+      }
+    };
+  } else {
+    allowedDeclaration = (declaration) =>
+      declaration.init?.type === 'CallExpression' &&
+      isRequireCall(declaration.init);
+  }
+
+  node.declarations
+    .filter((declaration) => !allowedDeclaration(declaration))
+    .forEach((declaration) => {
+      context.report({
+        node: declaration,
+        messageId: 'message'
+      });
+    });
 }
 
 export const noTopLevelVariables: Rule.RuleModule = {
@@ -47,10 +111,7 @@ export const noTopLevelVariables: Rule.RuleModule = {
     ]
   },
   create: (context) => {
-    const options: {
-      readonly constAllowed: ReadonlyArray<string>;
-      readonly kind: ReadonlyArray<string>;
-    } = {
+    const options: Options = {
       // type-coverage:ignore-next-line
       constAllowed: context.options[0]?.constAllowed || constAllowedValues,
       // type-coverage:ignore-next-line
@@ -58,54 +119,17 @@ export const noTopLevelVariables: Rule.RuleModule = {
     };
 
     return {
+      ExportNamedDeclaration: (node) => {
+        // type-coverage:ignore-next-line
+        const declaration = node.declaration as Declaration;
+        if (declaration.type === 'VariableDeclaration') {
+          checker(context, options, declaration);
+        }
+      },
       VariableDeclaration: (node) => {
-        if (
-          !isTopLevel(node) ||
-          !Array.from(options.kind).includes(node.kind)
-        ) {
-          return;
+        if (isTopLevel(node)) {
+          checker(context, options, node);
         }
-
-        let allowedDeclaration: (declaration: VariableDeclarator) => boolean;
-        if (node.kind === 'const') {
-          const isArrowFunctionAllowed = options.constAllowed.includes(
-            'ArrowFunctionExpression'
-          );
-          const isLiteralAllowed = options.constAllowed.includes('Literal');
-          const isMemberExpressionAllowed =
-            options.constAllowed.includes('MemberExpression');
-
-          allowedDeclaration = (declaration) => {
-            // type-coverage:ignore-next-line
-            switch ((declaration.init as Expression).type) {
-              case 'ArrowFunctionExpression':
-                return isArrowFunctionAllowed;
-              case 'CallExpression': {
-                // type-coverage:ignore-next-line
-                return isRequireCall(declaration.init as CallExpression);
-              }
-              case 'Literal':
-                return isLiteralAllowed;
-              case 'MemberExpression':
-                return isMemberExpressionAllowed;
-              default:
-                return false;
-            }
-          };
-        } else {
-          allowedDeclaration = (declaration) =>
-            declaration.init?.type === 'CallExpression' &&
-            isRequireCall(declaration.init);
-        }
-
-        node.declarations
-          .filter((declaration) => !allowedDeclaration(declaration))
-          .forEach((declaration) => {
-            context.report({
-              node: declaration,
-              messageId: 'message'
-            });
-          });
       }
     };
   }