+++ CHANGELOG.md
@@ -10,7 +10,14 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- (`ba81b6d`) Allow top-level `const` assignment from literals.
+- (`bc90491`) Optionally allow `Symbol()` for `no-top-level-side-effects`.
+- (`bed2d39`) Report named export declarations for `no-top-level-variables`.
+- (`efd1232`) Report function calls in exports for `no-top-level-side-effects`.
+- (`cd61a0a`) Report function calls in variable declarations for
+  `no-top-level-side-effects`.
+- (`fe11de9`) Report `new` calls in export and variable declarations for
+  `no-top-level-side-effects`.
 
 ## [2.1.0] - 2023-08-06